+++ stripe/src/test/java/com/stripe/android/model/ContextUtilsTest.kt
@@ -6,7 +6,7 @@ import android.content.pm.PackageManager
 import com.google.common.truth.Truth.assertThat
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
-import com.stripe.android.model.ContextUtils.packageInfo
+import com.stripe.android.utils.ContextUtils.packageInfo
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
